quiz-worker-py/pyproject.toml
[project]
name = "quiz-worker-py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "webtypy>=0.1.7",
    "pytz",
    "pyjwt",
]

[dependency-groups]
dev = [
    "workers-py",
]
